refactor: move getDraftDelay in its only usage
- rename store-utils.ts in store-selectors.ts
