+++ src/store/editor/hooks/updload.ts
@@ -5,10 +5,10 @@
  */
 import { useMemo } from 'react';
 
+import { selectUnsavedAttachmentByUploadId } from '../store-selectors';
 import { computeAndUpdateEditorStatus } from 'store/editor/hooks/commons';
 import { useSaveDraftFromEditor } from 'store/editor/hooks/save-draft';
 import { useEditorsStore } from 'store/editor/store';
-import { getUnsavedAttachmentIndex } from 'store/editor/store-utils';
 import { AttachmentUploadProcessStatus, MailsEditorV2 } from 'types/index.d';
 
 export const useEditorUploadProcess = (
@@ -17,15 +17,14 @@ export const useEditorUploadProcess = (
 ): { status: AttachmentUploadProcessStatus; cancel: () => void } | null => {
 	const { debouncedSaveDraft } = useSaveDraftFromEditor();
 	const attachmentStateInfo = useEditorsStore((state) => {
-		const unsavedAttachmentIndex = getUnsavedAttachmentIndex(state, editorId, uploadId);
-		if (unsavedAttachmentIndex === null) {
-			return null;
+		const unsavedAttachment = selectUnsavedAttachmentByUploadId(state, editorId, uploadId);
+		if (!unsavedAttachment) {
+			return unsavedAttachment;
 		}
 
 		return {
-			status: state.editors[editorId].unsavedAttachments[unsavedAttachmentIndex].uploadStatus,
-			abortController:
-				state.editors[editorId].unsavedAttachments[unsavedAttachmentIndex].uploadAbortController
+			status: unsavedAttachment.uploadStatus,
+			abortController: unsavedAttachment.uploadAbortController
 		};
 	});
